requirements.txt
numpy>=1.26
pygame-ce>=2.4
gymnasium>=0.29
pettingzoo>=1.24
torch>=2.2
scipy>=1.12
hypothesis>=6.100
pytest>=8.0
pytest-benchmark>=4.0
pyyaml>=6.0
matplotlib>=3.8
ruff>=0.4
